CB-490 fix: platform type error
플랫폼 타입을 불러오지 못하는 문제 수정
- usePlatform hook 삭제
- redux slice에서 로직 처리
